moved clean and replay to orchestrate
